Renamed 2 pngs
Renamed the UI_AeroplaneDamaged/2FinalDamage png and its import counterpart. I think any /'s are seen as colons (:) in file manager which is bad because windows does not accept colons or other special characters in file names, which I think is the reason why I can't download on windows. This should fix it?