Got scenes to load in the main game scene
Had to do some bug fixing to get here but it works. The GameScene scene loads a scene on startup and is currently set to only load the stopwatch example but should be configurable to load other scenes. You are able to interact with the sidebar and the instanced scene at the same time.