Re-organized files and made my own PrecisionMachiningGame
I reorganized files to make more sense. Still not done with it, but its a start. Also made my own Precision Machining minigame to substitute for Marshall's. Mine has 5 different shapes to cut out, instructions, a background, (sort of) a nozzle that has an animation, and only 1 script instead of 4 different ones. I also updated the title screen to have an image of the school and the proper game logo. After this I plan to implement more minigames.